1.处理业务部分重构，增加接口易用性和拓展性 2.修复增加第一笔业务时前端标签显示不符问题
