+++ accoj/blueprints/accoj.py
@@ -9,10 +9,11 @@ from flask import Blueprint, render_template, redirect, request, jsonify, sessio
 from accoj.utils import login_required
 from accoj.extensions import mongo
 from accoj.utils import is_number
-from _datetime import datetime
+from accoj.deal_business import deal_business
 import random
 
 accoj_bp = Blueprint('accoj', __name__)
+
 
 # 第一次课程----start-------------------------------------------------------------------------------
 @accoj_bp.route('/coursei', methods=['POST', 'GET'])
@@ -178,12 +179,12 @@ def revoke_add_business():
                 return jsonify(result=False, message="暂无业务")
             question_no = company.get("businesses")[-1].get("question_no")
             mongo.db.company.update(dict(_id=_id),
-                                    {"$pop": {"businesses": 1},
-                                     "$inc": {"business_num": -1},
+                                    {"$pop" : {"businesses": 1},
+                                     "$inc" : {"business_num": -1},
                                      "$pull": {"com_assets": {"question_no": question_no}}})
             mongo.db.company.update({"student_no": "{}_cp".format(session["username"])},
-                                    {"$pop": {"businesses": 1},
-                                     "$inc": {"business_num": -1},
+                                    {"$pop" : {"businesses": 1},
+                                     "$inc" : {"business_num": -1},
                                      "$pull": {"com_assets": {"question_no": question_no}}})
             return jsonify(result=True)
         else:
@@ -208,65 +209,24 @@ def add_business():
             return jsonify(result=False, message="业务类型错误！")
 
         if not company.get("businesses"):
-            # 刚开始增加业务,随机生成题库号
-            questions_no = random.randint(1, 1)
+            # 刚开始增加业务
+            questions_no_low = 1
+            questions_no_high = 1
+            # 随机生成题库号
+            questions_no = random.randint(questions_no_low, questions_no_high)
             # 第一笔业务，注册资本存入银行
-            question = mongo.db.question.find_one()
-            date = datetime.now()
-            year = date.year
-            month = date.month
-            day = 1
-            date = datetime(year, month, day)
-            content = question.get("content")
-            content = content.replace("v1", company.get("com_shareholder")[0])
-            num = company.get("com_regist_cap")
-            num *= 10000
-            content = content.replace("v2", str(int(num)))
-            business_type = question.get("business_type")
-            affect_type = question.get("affect_type")
-            values = [{"value_type": "common", "value": company.get("com_regist_cap")},
-                      {"value_type": "num", "value": num}]
-            key_element_infos = question.get("key_element_infos")
-            key_element_infos_len = len(key_element_infos)
-            for i in range(0, key_element_infos_len):
-                key_element_infos[i]["money"] = values[int(key_element_infos[i].get("value_index")) - 1].get("value")
-                key_element_infos[i].pop("value_index")
-            subjects_infos = question.get("subjects_infos")
-            subjects_infos_len = len(subjects_infos)
-            for i in range(0, subjects_infos_len):
-                subjects_infos[i]["money"] = values[int(subjects_infos[i].get("value_index")) - 1].get("value")
-                subjects_infos[i].pop("value_index")
-            content = "{}年{}月{}日，".format(year, month, day) + content
-            business_content = content
-            first_business = dict(questions_no=questions_no,
-                                  question_no=1,
-                                  content=content,
-                                  date=date,
-                                  business_type=business_type)
-            first_business_cp = dict(questions_no=questions_no,
-                                     question_no=1,
-                                     content=content,
-                                     date=date,
-                                     business_type=business_type,
-                                     affect_type=affect_type,
-                                     key_element_infos=key_element_infos,
-                                     subjects_infos=subjects_infos)
-            # 用户公司写入业务
-            mongo.db.company.update({"student_no": "{}".format(session["username"])},
-                                    {"$push": {"businesses": first_business},
-                                     "$inc": {"business_num": 1}})
-            # 副本公司存储答案，业务数增1
-            mongo.db.company.update({"student_no": "{}_cp".format(session["username"])},
-                                    {"$push": {"businesses": first_business_cp},
-                                     "$inc": {"business_num": 1}})
-            return jsonify(result=True, content=business_content)
+            if questions_no == 1:
+                business_content = deal_business(company=company, business_type="筹资业务", questions_no=1)
+                return jsonify(result=True, content=business_content)
 
         elif company.get("business_num") == 25:
             message = "业务数已达上限"
             return jsonify(result=False, message=message)
-        elif company["businesses"][0]["questions_no"] == 1:
-            # 公司已存在业务，且题库号为1
-            business_content, message = deal_business_1(company, business_type)
+        else:
+            # 公司已存在业务，获取题库号
+            questions_no = company["businesses"][0]["questions_no"]
+            business_content, message = deal_business(company=company, business_type=business_type,
+                                                      questions_no=questions_no)
             if not business_content:
                 return jsonify(result=False, message=message)
             else:
@@ -275,392 +235,6 @@ def add_business():
     return redirect('/coursei')
 
 
-def deal_business_1(company, com_business_type):
-    """
-    题库1的处理，返回业务内容，报错信息
-    :param company: company document
-    :param com_business_type: business_type
-    :return: business content,message
-    """
-    message = ""
-    com_businesses = company["businesses"]
-    business_num = company["business_num"]
-    add_question_content = None
-
-    if business_num == 24:
-        add_question_content = deal_with_question_1(company, 34)
-    else:
-        # 公司已有的业务
-        question_list = list()
-        for com_business in com_businesses:
-            question_list.append(com_business["question_no"])
-        # 筹资活动2-7
-        random_list = list(range(2, 8))
-        if com_business_type == "投资活动":
-            random_list = list(range(8, 15))
-        elif com_business_type == "经营活动":
-            random_list = list(range(15, 34))
-        for question_no in question_list:
-            if question_no in random_list:
-                random_list.remove(question_no)
-        print("question_list:{}".format(question_list))
-        print("random_list:{}".format(random_list))
-        if not random_list:
-            message = "请选择其他活动"
-            return add_question_content, message
-        while True:
-            # 随机选择题目
-            if not random_list:
-                # 随机数序列为空
-                message = "请选择其他活动"
-                print(message)
-                return add_question_content, message
-            add_question_no = random.choice(random_list)
-            if add_question_no in question_list:
-                continue
-            elif add_question_no == 3 and 2 not in question_list:
-                # 业务2为业务3前提
-                random_list.remove(3)
-                continue
-            elif add_question_no == 13 and 12 not in question_list:
-                # 业务12为业务13前提
-                random_list.remove(13)
-                continue
-            elif add_question_no == 18 and 17 not in question_list:
-                # 业务17为业务18前提
-                random_list.remove(18)
-                continue
-            elif add_question_no == 20 and 15 not in question_list:
-                # 业务15为业务20前提
-                random_list.remove(15)
-                continue
-            elif add_question_no == 23 and 19 not in question_list:
-                # 业务19为业务23前提
-                random_list.remove(23)
-                continue
-            elif add_question_no == 27 and 5 not in question_list and 8 not in question_list and 9 not in question_list:
-                # 业务[5,8,8]为业务27前提
-                random_list.remove(27)
-                continue
-            elif add_question_no == 28 and 24 not in question_list and 25 not in question_list and \
-                    26 not in question_list and 27 not in question_list:
-                # 业务[24,45,26,27]为业务28前提
-                random_list.remove(28)
-                continue
-            elif add_question_no == 30 and 28 not in question_list:
-                # 业务28为业务30前提
-                random_list.remove(30)
-                continue
-            elif add_question_no == 31 and 28 not in question_list and 30 not in question_list:
-                # 业务28为业务31前提
-                random_list.remove(31)
-                continue
-            elif add_question_no == 33 and 25 not in question_list:
-                # 业务25为业务33前提
-                random_list.remove(33)
-                continue
-            else:
-                add_question_content = deal_with_question_1(company, add_question_no)
-                break
-
-    return add_question_content, message
-
-
-def deal_with_question_1(company, question_no):
-    """
-    对应题号生成对应业务
-    :param company: company document
-    :param question_no: question no
-    :return: business content
-    """
-    question = mongo.db.question.find_one(dict(questions_no=1,
-                                               question_no=question_no))
-    businesses = company.get("businesses")
-    com_assets = company["com_assets"]
-    date = datetime.now()
-    year = date.year
-    month = date.month
-    day = company.get("business_num") + 1
-    if question_no == 34 or question_no == 33:
-        # 当前月份天数即当月最后一天
-        day = (datetime(year, month + 1, 1) - datetime(year, month, 1)).days
-    date = datetime(year, month, day)
-    content = question.get("content")
-    business_type = question.get("business_type")
-    affect_type = question.get("affect_type")
-    key_element_infos = question.get("key_element_infos")
-    subjects_infos = question.get("subjects_infos")
-
-    # deal values and content------------------------------------------------------------------------
-    values = question.get("values")
-    values_len = len(values)
-    values_list = list()
-    for i in range(0, values_len):
-        value = values[i].get("value")
-        value_type = values[i].get("value_type")
-        is_random = values[i].get("is_random")
-        low = values[i].get("low")
-        high = values[i].get("high")
-        if question_no == 13 and i == 3:
-            # 问题13特判
-            value_tmp = 0
-            for asset in com_assets:
-                if asset.get("asset_name") == values_list[0]:
-                    value_tmp = values_list[1] - asset.get("asset_value")
-                    break
-            if values_list[1] >= value_tmp:
-                value = value.split("/")[0]
-            else:
-                value = value.split("/")[1]
-                # 收益为负
-                subjects_infos[i]["is_up"] = False
-        elif question_no == 23 and (i == 2 or i == 3):
-            # 问题23特判
-            for business in businesses:
-                if business.get("question_no") == 20:
-                    value_tmp = business.get("key_element_infos")[0].get("money")
-                    if values_list[1] > value_tmp:
-                        value = value.split("/")[0]
-                    elif values_list[1] < value_tmp:
-                        value = value.split("/")[1]
-                    else:
-                        if i == 2:
-                            content_tmp = content.split("，")
-                            content_tmp.pop(-1)
-                            content = "，".join(content_tmp)
-                    break
-        elif value_type == "asset":
-            if value[0] == "*":
-                value = value.lstrip("*")
-            elif value[0] == "-":
-                value = value.lstrip("-")
-                for asset in com_assets:
-                    if question_no == 30:
-                        # 问题30特判
-                        break
-                    if asset.get("asset_name") == value:
-                        com_assets.remove(asset)
-                        break
-        elif value_type == "num":
-            if is_random:
-                value = random.randrange(low, high, 100)
-            if i and question_no != 28 and values[i - 1].get("value_type") == "asset" and \
-                    values[i - 1].get("value")[0] == "+":
-                # 公司资产增加处理
-                content = content.replace("+", "")
-                asset_name = values_list[i - 1].replace("+", "")
-                com_assets.append(dict(asset_name=asset_name, market_value=value, question_no=question_no))
-            if question_no == 2 and i == 0:
-                value = random.randint(low, high)
-            elif question_no == 13 and i == 2:
-                # 问题13特判
-                for asset in com_assets:
-                    if asset.get("asset_name") == values_list[0]:
-                        value = abs(values_list[1] - asset.get("asset_value"))
-                        break
-            elif question_no == 14:
-                # 问题14特判
-                value = 10000 * values_list[1]
-            elif question_no == 18:
-                # 问题18特判
-                for business in businesses:
-                    if business.get("question_no") == 17:
-                        high = business.get("key_element_infos")[0].get("money")
-                        value = random.randrange(90000, high, 100)
-                        break
-            elif question_no == 20:
-                # 问题20特判
-                for business in businesses:
-                    if business.get("question_no") == 15:
-                        high = business.get("key_element_infos")[0].get("money")
-                        value = random.randrange(1000, high, 100)
-                        break
-            elif question_no == 23:
-                # 问题23特判
-                for business in businesses:
-                    if business.get("question_no") == 20:
-                        medium = business.get("key_element_infos")[0].get("money")
-                        value = random.randrange(medium - 1000, medium + 1000, 100)
-                        break
-            elif question_no == 27:
-                # 问题27特判
-                pass
-            elif question_no == 28:
-                # 问题28特判
-                content = content.replace("+", "")
-                asset_name = values_list[0].replace("+", "")
-                com_assets.append(dict(asset_name=asset_name, market_value=value))
-                for business in businesses:
-                    value = 0
-                    if business.get("question_no") in [24, 25, 26, 27]:
-                        value += business.get("subjects_infos")[0].get("money")
-            elif question_no == 30:
-                # 问题30特判
-                if i == 2:
-                    for business in businesses:
-                        if business.get("question_no") == 28:
-                            tmp = business.get("key_element_infos")[0].get("money")
-                            value = tmp / 2
-                            break
-                if i == 3:
-                    value = int(values_list[i - 1] * 0.8)
-            elif question_no == 31:
-                # 问题31特判
-                if i == 2:
-                    for business in businesses:
-                        if business.get("question_no") == 30:
-                            value = business.get("key_element_infos")[0].get("money")
-                            break
-                if i == 3:
-                    value = int(values_list[i - 1] * 0.8)
-            elif question_no == 33:
-                # 问题33特判
-                if i == 2:
-                    for business in businesses:
-                        if business.get("question_no") == 28:
-                            tmp = business.get("key_element_infos")[2].get("money")
-                            value = tmp / 2
-                            break
-                if i == 3:
-                    value = int(values_list[i - 1] * 0.8)
-
-        elif value_type == "percent":
-            if is_random:
-                value = random.randint(low, high)
-        values_list.append(value)
-        if value_type == "percent":
-            content = content.replace("v{}".format(i + 1), str(value) + "%")
-        else:
-            content = content.replace("v{}".format(i + 1), str(value))
-    # end---------------------------------------------------------------------------
-
-    # deal key_element_infos--------------------------------------------------------
-    key_element_infos_len = len(key_element_infos)
-    for i in range(0, key_element_infos_len):
-        value_index = key_element_infos[i].get("value_index")
-        if value_index[0] != "(":
-            money = values_list[int(value_index) - 1]
-        else:
-            value_index_len = len(value_index)
-            money = values_list[int(value_index[1]) - 1]
-            symbol = True
-            for j in range(2, value_index_len - 1):
-                if j % 2 == 0:
-                    if value_index[j] == "+":
-                        symbol = True
-                    else:
-                        symbol = False
-                else:
-                    if symbol:
-                        money += values_list[int(value_index[j]) - 1]
-                    else:
-                        money -= values_list[int(value_index[1]) - 1]
-        key_element_infos[i]["money"] = money
-        key_element_infos[i].pop("value_index")
-    # end---------------------------------------------------------------------------
-
-    # deal subjects_infos-----------------------------------------------------------
-    subjects_infos_len = len(subjects_infos)
-    for i in range(0, subjects_infos_len):
-        value_index = subjects_infos[i].get("value_index")
-        if (i and question_no == 2) or (not i and question_no == 3):
-            content_tmp = None
-            if question_no == 2:
-                # 问题2特判
-                year_tmp = int(content.find("年") - 1)
-            else:
-                # 问题3特判
-                for business in businesses:
-                    if business.get("question_no") == 2:
-                        content_tmp = business.get("content")
-                        break
-                year_tmp = int(content_tmp.find("年") - 1)
-            subject_temp = subjects_infos[i].get("subject")
-            if year_tmp == 1:
-                subjects_infos[i]["subject"] = subject_temp.split("/")[0]
-            else:
-                if question_no == 2:
-                    subjects_infos[i]["subject"] = subject_temp.split("/")[1]
-        if question_no == 23:
-            # 问题23特判
-            money = None
-            if content.find("多于"):
-                value_tmp = 0
-                for business in businesses:
-                    if business.get("question_no") == 20:
-                        value_tmp = business.get("key_element_infos")[0].get("money")
-                        break
-                if i == 0:
-                    money = values_list[0]
-                elif i == 1:
-                    money = value_tmp
-                else:
-                    subjects_infos.append(dict(subject="银行存款", money=values_list[0] - value_tmp, is_up=False))
-            elif content.find("少于"):
-                if i == 0:
-                    money = values_list[0]
-                elif i == 1:
-                    money = values_list[0]
-                else:
-                    value_tmp = 0
-                    for business in businesses:
-                        if business.get("question_no") == 20:
-                            value_tmp = business.get("key_element_infos")[0].get("money")
-                            break
-                    subjects_infos.append(dict(subject="银行存款",
-                                               money=abs(values_list[0] - value_tmp),
-                                               is_up=False))
-            else:
-                money = values_list[int(value_index) - 1]
-        elif value_index[0] != "(":
-            money = values_list[int(value_index) - 1]
-        else:
-            value_index_len = len(value_index)
-            money = values_list[int(value_index[1]) - 1]
-            symbol = True
-            for j in range(2, value_index_len - 1):
-                if j % 2 == 0:
-                    if value_index[j] == "+":
-                        symbol = True
-                    else:
-                        symbol = False
-                else:
-                    if symbol:
-                        money += values_list[int(value_index[j]) - 1]
-                    else:
-                        money -= values_list[int(value_index[1]) - 1]
-        if money:
-            subjects_infos[i]["money"] = money
-        subjects_infos[i].pop("value_index")
-    # end---------------------------------------------------------------------------
-
-    content = "{}年{}月{}日，".format(year, month, day) + content
-    business = dict(questions_no=1,
-                    question_no=question_no,
-                    content=content,
-                    date=date,
-                    business_type=business_type)
-
-    business_cp = dict(questions_no=1,
-                       question_no=question_no,
-                       content=content,
-                       date=date,
-                       business_type=business_type,
-                       affect_type=affect_type,
-                       key_element_infos=key_element_infos,
-                       subjects_infos=subjects_infos)
-    # 用户公司写入业务
-    mongo.db.company.update({"student_no": "{}".format(session["username"])},
-                            {"$push": {"businesses": business},
-                             "$set": {"com_assets": com_assets},
-                             "$inc": {"business_num": 1}})
-    # 副本公司存储答案
-    mongo.db.company.update({"student_no": "{}_cp".format(session["username"])},
-                            {"$push": {"businesses": business_cp},
-                             "$set": {"com_assets": com_assets},
-                             "$inc": {"business_num": 1}}
-                            )
-    return content
 # 第一次课程----end---------------------------------------------------------------------------------
 
 
@@ -672,6 +246,8 @@ def courseii():
     :return:
     """
     return render_template('course/courseii.html')
+
+
 # 第二次课程----end---------------------------------------------------------------------------------
 
 
@@ -683,6 +259,8 @@ def courseiii():
     :return:
     """
     return render_template('course/courseiii.html')
+
+
 # 第三次课程----end---------------------------------------------------------------------------------
 
 
@@ -694,6 +272,8 @@ def courseiv():
     :return:
     """
     return render_template('course/courseiv.html')
+
+
 # 第四次课程----end---------------------------------------------------------------------------------
 
 
@@ -714,6 +294,8 @@ def coursev_2():
     :return:
     """
     return render_template('course/coursev_2.html')
+
+
 # 第五次课程----end---------------------------------------------------------------------------------
 
 
@@ -725,6 +307,8 @@ def coursevi():
     :return:
     """
     return render_template('course/coursevi.html')
+
+
 # 第六次课程----end---------------------------------------------------------------------------------
 
 
@@ -736,6 +320,8 @@ def coursevii():
     :return:
     """
     return render_template('course/coursevii.html')
+
+
 # 第七次课程----end---------------------------------------------------------------------------------
 
 
@@ -747,6 +333,8 @@ def courseviii():
     :return:
     """
     return render_template('course/courseviii.html')
+
+
 # 第八次课程----end---------------------------------------------------------------------------------
 
 
@@ -785,6 +373,8 @@ def courseix_4():
     :return:
     """
     return render_template('course/courseix_4.html')
+
+
 # 第九次课程----end---------------------------------------------------------------------------------
 
 
@@ -796,6 +386,8 @@ def coursex():
     :return:
     """
     return render_template('course/coursex.html')
+
+
 # 第十次课程----end---------------------------------------------------------------------------------
 
 
@@ -807,6 +399,8 @@ def detail():
     :return:
     """
     return render_template('detail.html')
+
+
 # 用户个人中心----end-------------------------------------------------------------------------------
